edit type data for struct item

+++ model/webhook.go
@@ -109,7 +109,7 @@ type MenuItemtes struct {
 	Name          string  `json:"name" bson:"name"`
 	Price         int     `json:"price" bson:"price"`
 	OriginalPrice int     `json:"originalPrice" bson:"originalPrice"`
-	Rating        float32 `json:"rating" bson:"rating"`
+	Rating        float64 `json:"rating" bson:"rating"`
 	Sold          string  `json:"sold" bson:"sold"`
 	Image         string  `json:"image" bson:"image"`
 }
